Get current configuration (#3)
Until now we only support registering a callback to receive the new configuration when it's updated.
Now we also support getting the current configuration, which is simpler.
Get is important for the dynamic and static configurations to be in the same file.

In the future, the manager may know whether the configuration changed anywhere in a subtree where there's no one registered, and this will be a sign that a restart is required (because if there's no one registered then the configuration obviously can't be altered dynamically).